Consolidate amount utilities to fix CI issues
- Update all imports from @/utils/amount to @/lib/amount
- Fix function calls to use formatAmount instead of fromStroops where appropriate
- Remove duplicate utils/amount.ts file
- Fix duplicate import in dashboard-view.tsx
- Resolves TypeScript compilation errors related to amount formatting
