+++ frontend/src/components/stream-creation/AmountStep.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useRef, useEffect } from "react";
-import { hasValidPrecision, toStroops } from "@/utils/amount";
+import { hasValidPrecision, toStroops } from "@/lib/amount";
 
 interface AmountStepProps {
   value: string;
